requirement.txt
flask
flask_sqlalchemy
flask_bcrypt
flask_login
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
kafka-python==2.0.2
